2월20일 수업복습 과제 수행완료

+++ src/main/java/com/sbs/untact/controller/UsrArticleController.java
@@ -3,6 +3,8 @@ package com.sbs.untact.controller;
 import java.util.List;
 import java.util.Map;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -10,8 +12,11 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.sbs.untact.dto.Article;
+import com.sbs.untact.dto.Board;
+import com.sbs.untact.dto.Reply;
 import com.sbs.untact.dto.ResultData;
 import com.sbs.untact.service.ArticleService;
+import com.sbs.untact.util.Util;
 
 @Controller
 public class UsrArticleController {
@@ -20,15 +25,30 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/detail")
 	@ResponseBody
-	public Article showDetail(int id) {
-		Article article = articleService.getArticle(id);
+	public ResultData showDetail(Integer id) {
+		if (id == null) {
+			return new ResultData("F-1", "id를 입력해주세요.");
+		}
 
-		return article;
+		Article article = articleService.getForPrintArticle(id);
+
+		if (article == null) {
+			return new ResultData("F-1", "해당 게시물은 존재하지 않습니다.");
+		}
+
+		return new ResultData("S-1", "성공 입니다.", "aritcle", article);
+
 	}
 
 	@RequestMapping("/usr/article/list")
 	@ResponseBody
-	public List<Article> showList(String searchKeywordType, String searchKeyword) {
+	public ResultData showList(@RequestParam(defaultValue = "1") int boardId, String searchKeywordType, String searchKeyword, @RequestParam(defaultValue = "1") int page) {
+		
+		Board foundBoard = articleService.getBoard(boardId);
+		if(foundBoard == null) {
+			return new ResultData("F-2", "존재하지않는 게시판입니다.");
+		}
+		
 		if (searchKeywordType != null) {
 			searchKeywordType = searchKeywordType.trim();
 		}
@@ -47,27 +67,55 @@ public class UsrArticleController {
 		if (searchKeyword == null) {
 			searchKeywordType = null;
 		}
+		int itemsAPage = 10;
+		
+		List<Article> articles = articleService.getForPrintArticles(boardId, searchKeywordType, searchKeyword, itemsAPage, page);
 
-		return articleService.getArticles(searchKeywordType, searchKeyword);
+		return new ResultData("S-1", "성공", "articles", articles);
 	}
 
 	@RequestMapping("/usr/article/doAdd")
 	@ResponseBody
-	public ResultData doAdd(@RequestParam Map<String, Object> param) {
+	public ResultData doAdd(@RequestParam Map<String, Object> param, HttpSession session) {
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+
 		if (param.get("title") == null) {
 			return new ResultData("F-1", "title을 입력해주세요.");
 		}
 
-		if (param.get("title") == null) {
+		if (param.get("body") == null) {
 			return new ResultData("F-1", "body를 입력해주세요.");
 		}
 
+		param.put("memberId", loginedMemberId);
+
 		return articleService.addArticle(param);
 	}
+	
+	@RequestMapping("/usr/article/doAddReply")
+	@ResponseBody
+	public ResultData doAddReply(@RequestParam Map<String, Object> param, HttpSession session) {
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+
+		if (param.get("articleId") == null) {
+			return new ResultData("F-1", "articleId을 입력해주세요.");
+		}
+
+		if (param.get("body") == null) {
+			return new ResultData("F-1", "body를 입력해주세요.");
+		}
+
+		param.put("memberId", loginedMemberId);
+
+		return articleService.addReply(param);
+	}
+	
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(Integer id) {
+	public ResultData doDelete(Integer id, HttpSession session) {
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+
 		if (id == null) {
 			return new ResultData("F-1", "id를 입력해주세요.");
 		}
@@ -78,12 +126,20 @@ public class UsrArticleController {
 			return new ResultData("F-1", "해당 게시물은 존재하지 않습니다.");
 		}
 
+		ResultData actorCanDeleteRd = articleService.getActorCanDeleteRd(article, loginedMemberId);
+
+		if (actorCanDeleteRd.isFail()) {
+			return actorCanDeleteRd;
+		}
+
 		return articleService.deleteArticle(id);
 	}
 
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(Integer id, String title, String body) {
+	public ResultData doModify(Integer id, String title, String body, HttpSession session) {
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+
 		if (id == null) {
 			return new ResultData("F-1", "id를 입력해주세요.");
 		}
@@ -102,6 +158,94 @@ public class UsrArticleController {
 			return new ResultData("F-1", "해당 게시물은 존재하지 않습니다.");
 		}
 
+		ResultData actorCanModifyRd = articleService.getActorCanModifyRd(article, loginedMemberId);
+
+		if (actorCanModifyRd.isFail()) {
+			return actorCanModifyRd;
+		}
+
 		return articleService.modifyArticle(id, title, body);
 	}
+	
+	@RequestMapping("/usr/article/doModifyReply")
+	@ResponseBody
+	public ResultData doModifyReply(Integer id,Integer articleId, String body, HttpSession session) {
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+
+		if (id == null) {
+			return new ResultData("F-1", "id를 입력해주세요.");
+		}
+		
+		if (articleId == null) {
+			return new ResultData("F-1", "articleid를 입력해주세요.");
+		}
+
+		if (body == null) {
+			return new ResultData("F-1", "body를 입력해주세요.");
+		}
+
+		Article article = articleService.getArticle(articleId);
+
+		if (article == null) {
+			return new ResultData("F-1", "해당 게시물은 존재하지 않습니다.");
+		}
+		
+		Reply reply = articleService.getReply(id);
+
+		if (reply == null) {
+			return new ResultData("F-1", "해당 게시물의 댓글은 없습니다.");
+		}
+		
+		ResultData actorCanModifyReplyRd = articleService.getActorCanModifyReplyRd(reply, loginedMemberId);
+
+		if (actorCanModifyReplyRd.isFail()) {
+			return actorCanModifyReplyRd;
+		}
+
+		return articleService.modifyReply(id, body);
+	}
+	
+	@RequestMapping("/usr/article/doDeleteReply")
+	@ResponseBody
+	public ResultData doDeleteReply(Integer id,int articleId , HttpSession session) {
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+
+		if (id == null) {
+			return new ResultData("F-1", "id를 입력해주세요.");
+		}
+
+		Article article = articleService.getArticle(articleId);
+
+		if (article == null) {
+			return new ResultData("F-1", "해당 게시물은 존재하지 않습니다.");
+		}
+		
+		Reply reply = articleService.getReply(id);
+
+		if (reply == null) {
+			return new ResultData("F-1", "해당 게시물의 댓글은 없습니다.");
+		}
+
+		ResultData actorCanDeleteReplyRd = articleService.getActorCanDeleteReplyRd(reply, loginedMemberId);
+
+		if (actorCanDeleteReplyRd.isFail()) {
+			return actorCanDeleteReplyRd;
+		}
+
+		return articleService.deleteReply(id);
+	}
+	
+	@RequestMapping("/usr/article/replies")
+	@ResponseBody
+	public ResultData replies(Integer articleId) {
+		
+		if(articleId == null) {
+			return new ResultData("F-1", "articleId를 입력해주세요");
+		}
+		
+		List<Reply> replies = articleService.getForPrintReplies(articleId);
+
+		return new ResultData("S-1", "성공", "replies", replies);
+	}
+	
 }
